\name{rolling}
\alias{rolling}
\docType{data}
\title{Rolling Thumbtacks Data}
\description{
 This data was generated by Beckett and Diaconis (1994). They generate
 binary strings from rools of common thumbtacks. A 1 was recorded if the tack landed
 point up and 0 was recorder if the tack landed point down. All tacks started point down.
 Each tack was flicked 9 times. The data consit of 320 9-tuples. The actual data arose 
 from 16 "flickers" and 10 surfaces. Following Liu (1996), we treat the data as though
 they came from 320 different tacks. 
}
\usage{data(rolling)}
\format{
  A data frame with 320 observations on the following 2 variables.
  \describe{
    \item{\code{y1}}{a numeric vector giving the number of tacks landed point up.}
    \item{\code{y2}}{a numeric vector giving the number of trials.}
  }
}
\source{
Beckett, L. and Diaconis. P. (1994). Spectral analysis fro discrete longitudinal 
  data. Adv. Math., 103: 107-128.
}
\references{
Liu, J.S. (1996). Nonparametric Hierarchical Bayes via Sequential Imputations. 
  The Annals of Statistics, 24: 911-930.
}
\examples{
data(rolling)
}
\keyword{datasets}
